requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
groq==0.11.0
openai==1.51.0
numpy==1.26.4
sentence-transformers==3.1.1
